@@ -12,17 +12,17 @@
 	},
 	"homepage": "https://github.com/timschwab/clerk",
 	"dependencies": {
-		"axios": "^0.24.0",
+		"axios": "^0.27.2",
 		"express": "^4.17.1",
 		"uuid": "^8.3.2",
-		"vue": "^2.6.14",
-		"vue-router": "^3.5.3",
-		"vuex": "^3.6.2"
+		"vue": "^3.2.37",
+		"vue-router": "^4.1.4",
+		"vuex": "^4.0.2"
 	},
 	"devDependencies": {
-		"lint-staged": "^12.0.1",
+		"lint-staged": "^13.0.3",
 		"prettier": "^2.4.1",
-		"vue-loader": "^15.9.8",
+		"vue-loader": "^17.0.0",
 		"vue-style-loader": "^4.1.3",
 		"vue-template-compiler": "^2.6.14",
 		"webpack": "^5.64.0",
