@@ -1,39 +1,47 @@
-// Set up vue
-import Vue from 'vue';
-import VueRouter from 'vue-router';
-
-import HomePage from './HomePage.vue';
-import NotFoundPage from './404.vue';
-import RandomFunds from './RandomFunds.vue';
-import Christmas from './Christmas.vue';
-
-Vue.use(VueRouter);
+import { createRouter, createWebHistory } from 'vue-router';
 
 // Includes
+import HomePage from './HomePage.vue';
+import NotFoundPage from './404.vue';
+import LoginPage from './LoginPage.vue';
+// import RandomFunds from './RandomFunds.vue';
+// import Christmas from './Christmas.vue';
+
 let store; // Injected by app.js
 
 // Base pages
-const routing = new VueRouter({
-	mode: 'history',
+const routing = createRouter({
+	history: createWebHistory(),
 	routes: [
 		{
 			path: '/',
+			name: 'home',
 			component: HomePage
 		},
 		{
-			path: '/random-funds',
-			component: RandomFunds
+			path: '/login',
+			name: 'login',
+			component: LoginPage
 		},
 		{
-			path: '/christmas',
-			component: Christmas
-		},
-		{
-			path: '*',
+			path: '/:pathMatch(.*)*',
+			name: '404',
 			component: NotFoundPage
 		}
 	]
 });
+
+///*
+routing.beforeEach((to, from) => {
+	if (store.getters.authenticated || to.name == 'login') {
+		return true;
+	} else {
+		return {
+			name: 'login'
+		};
+	}
+});
+//*/
 
 if (DEV_MODE) {
 	// Expose globally for debugging
