@@ -1,3 +1,9 @@
 export default {
-	//
+	authenticated(state) {
+		if (state.token) {
+			return true;
+		} else {
+			return false;
+		}
+	}
 }
